pyproject.toml
[project]
name = "ral"
version = "0.1.0"
description = "练手项目"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentskills-fs>=0.1.0",
    "agentskills-langchain>=0.2.0",
    "dashscope>=1.25.12",
    "fastapi>=0.134.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "llama-index>=0.14.15",
    "llama-index-embeddings-langchain>=0.4.1",
    "llama-index-llms-langchain>=0.7.2",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "qdrant-client>=1.17.0",
    "sqlmodel>=0.0.14",
    "python-multipart>=0.0.6",
    "pypdf>=3.0.0",
    "python-docx>=1.0.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "langchain-mcp-adapters>=0.2.1",
]
